requirements.txt
appdirs==1.4.4
appnope==0.1.2
asgiref==3.3.4
attrs==21.2.0
backcall==0.2.0
black==19.10b0
Brotli==1.0.9
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
decorator==5.0.9
Django==3.2.4
django-cors-headers==3.7.0
django-extensions==3.1.3
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
httprunner==3.1.5
idna==2.10
ipython==7.24.1
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==2.11.3
jmespath==0.9.5
loguru==0.4.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
more-itertools==8.8.0
mysqlclient==2.0.3
packaging==20.9
parso==0.8.2
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.19
ptyprocess==0.7.0
py==1.10.0
pydantic==1.8.2
Pygments==2.9.0
PyJWT==1.7.1
PyMySQL==1.0.2
pyparsing==2.4.7
pytest==5.4.3
pytest-html==2.1.1
pytest-metadata==1.11.0
pytz==2021.1
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
sentry-sdk==0.14.4
sqlparse==0.4.1
toml==0.10.2
traitlets==5.0.5
typed-ast==1.4.3
typing-extensions==3.10.0.0
urllib3==1.26.6
wcwidth==0.2.5
drf-yasg